requirements.txt
streamlit==1.28.1
PyPDF2
cohere
numpy
plotly
python-dotenv
